HACK: Add back static symbols and _jet header
The jemalloc 5 build system uses a combination of 'nm' and shell
scripts and 'awk' to generate a symbol list based on symbols present
in the compiled objects themselves.

This is a bit of a hassle at this point to develop all that
infrastructure inside our CMake build system.

Longer-term, we'd need to:

- Convert looping text manipulation shell scripts in
include/jemalloc/internal/*.sh to CMake text manipulation functions.
- Add custom build targets to run 'nm' against our intermediate objects.
- Send the output of 'nm' to something that will generate the proper
headers and symbol files.

For now, it's easier just copying the two files we need to keep the
build working into our infrastructure.